+- add patches.fixes/linux-post-2.6.2-20040205 +  altix update, qla update, many other fixes
suse-commit: b96c40e12cbcb5f39b6775656e65481784f69c9a

+++ include/asm-i386/hw_irq.h
@@ -26,7 +26,7 @@
  */
 
 extern u8 irq_vector[NR_IRQ_VECTORS];
-#define IO_APIC_VECTOR(irq)	((int)irq_vector[irq])
+#define IO_APIC_VECTOR(irq)	(irq_vector[irq])
 
 extern void (*interrupt[NR_IRQS])(void);
